% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_counties.R
\docType{data}
\name{census_counties}
\alias{census_counties}
\title{NJ Counties from the US Census}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 1 columns.
}
\usage{
census_counties
}
\description{
This is simply a list of all NJ counties from the US Census
}
\keyword{datasets}
