% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_municipalities.R
\docType{data}
\name{census_municipalities}
\alias{census_municipalities}
\title{NJ Municipalities from the US Census}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 565 rows and 4 columns.
}
\usage{
census_municipalities
}
\description{
This is a table all NJ municipalities from the US Census.
}
\keyword{datasets}
