% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident_location_type.R
\docType{data}
\name{incident_location_type}
\alias{incident_location_type}
\title{Location type}
\format{
A dataframe with 2 columns
\describe{
\item{form_id}{Unique identifier for the \code{incident} table.}
\item{location_type}{Incident location type.}
}
}
\source{
\url{https://www.njoag.gov/force/}
}
\usage{
incident_location_type
}
\description{
Location type
}
\details{
This data is extracted from the \code{location_type} field in
\code{use_of_force_raw}.

There may be multiple rows in this table linking to a single incident.
}
\seealso{
\url{https://nj.gov/oag/excellence/docs/Use-of-Force-Reporting-Portal-Guide.pdf}
}
\keyword{datasets}
