% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident_officer_medical_treatment.R
\docType{data}
\name{incident_officer_medical_treatment}
\alias{incident_officer_medical_treatment}
\title{Type of officer medical treatment.}
\format{
A dataframe with 2 columns
\describe{
\item{form_id}{Unique identifier for the \code{incident} table.}
\item{officer_medical_treatment}{Officer medical treament type. See details.}
}
}
\source{
\url{https://www.njoag.gov/force/}
}
\usage{
incident_officer_medical_treatment
}
\description{
Type of medical treatment received by officer.
}
\details{
This data is extracted from the \code{OFFMEDTREAT2} field in
\code{use_of_force_raw}.

There may be multiple rows in this table linking to a single incident.

The injury type will be one of the following:
\itemize{
\item "EMS on scene": Treated by Emergency Medical Services on scene.
\item "Hospital": Officer transported to hospital for care.
\item "Officer Administered First Aid": First aid administered by a police officer including cleaning cuts, applying antibiotic cream, bandages, or ice.
\item "Refused": Officer refused medical treatment.
\item "Urgent Care": Treated at an Urgent Care Facility
}
}
\seealso{
\url{https://nj.gov/oag/excellence/docs/Use-of-Force-Reporting-Portal-Guide.pdf}
}
\keyword{datasets}
