% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident_planned_contact.R
\docType{data}
\name{incident_planned_contact}
\alias{incident_planned_contact}
\title{Pre-planned contact.}
\format{
A dataframe with 2 columns
\describe{
\item{form_id}{Unique identifier for the \code{incident} table.}
\item{planned_contact}{Type of planned contact. See details.}
}
}
\source{
\url{https://www.njoag.gov/force/}
}
\usage{
incident_planned_contact
}
\description{
The type of pre-planned contact that initiated the police-citizen contact.
}
\details{
This data is extracted from the \code{planned_contact} field in
\code{use_of_force_raw}.

There may be multiple rows in this table linking to a single incident.

The planned contact will be one of the following:
\itemize{
\item "Arrest": An arrest of an individual executed at pre-planned time.
\item "Prisoner Transfer": Transport or transfer of a person in custody planned prior to execution.
\item "Processing": Processing of a person in custody planned prior to execution.
\item "Search Warrant Execution": Execution of an issued search warrant.
}
}
\seealso{
\url{https://nj.gov/oag/excellence/docs/Use-of-Force-Reporting-Portal-Guide.pdf}
}
\keyword{datasets}
