% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident_type.R
\docType{data}
\name{incident_type}
\alias{incident_type}
\title{Reason for Initial Dispatch or Encounter}
\format{
A dataframe with 2 columns
\describe{
\item{form_id}{Unique identifier for the \code{incident} table.}
\item{type}{Incident type.}
}
}
\source{
\url{https://www.njoag.gov/force/}
}
\usage{
incident_type
}
\description{
The type of incident during which the use of force occurred.
}
\details{
This data is extracted from the \code{incident_type} field in
\code{use_of_force_raw}.

There may be multiple rows in this table linking to a single incident.
}
\seealso{
\url{https://nj.gov/oag/excellence/docs/Use-of-Force-Reporting-Portal-Guide.pdf}
}
\keyword{datasets}
