% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident_video_type.R
\docType{data}
\name{incident_video_type}
\alias{incident_video_type}
\title{Video footage type.}
\format{
A dataframe with 2 columns
\describe{
\item{form_id}{Unique identifier for the \code{incident} table.}
\item{video_type}{Video type. See details.}
}
}
\source{
\url{https://www.njoag.gov/force/}
}
\usage{
incident_video_type
}
\description{
The type of video footage of the incident
}
\details{
This data is extracted from the \code{video_type} field in
\code{use_of_force_raw}.

There may be multiple rows in this table linking to a single incident.

The video type will be one of the following:
\itemize{
\item "Body Worn": Video footage from a body worn camera worn by any officer(s) or other emergency personnel on scene at the time of the use of force incident.
\item "Cell Phone": Video footage from a cell phone, tablet or similar device of any individual at the scene including witnesses or subjects.
\item "CED Camera": Video footage from a CED.
\item "Commercial Building": Video footage from a commercial location, including either exterior building or interior surveillance cameras.
\item "Motor Vehicle": Video footage from a motor vehicle recording system such as an MVR or DIVR and including citizen motor vehicle cameras.
\item "Residential/Home": Video footage from a private residential device, such as a Ring or Blink system or any other private recording equipment in the area around the incident, including CCTVs.
\item "Station House": Video footage from inside the police station
\item "Unknown": The source of the recording is unknown
\item "Other"
}
}
\seealso{
\url{https://nj.gov/oag/excellence/docs/Use-of-Force-Reporting-Portal-Guide.pdf}
}
\keyword{datasets}
