% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njoajuof-package.R
\docType{package}
\name{njoaguof-package}
\alias{njoaguof}
\alias{njoaguof-package}
\title{njoaguof: NJ OAG Use of Force data package}
\description{
This is cleaned-up version of the NJ OAG Use of Force database available from https://www.njoag.gov/force/.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tor-gu/njoaguof}
  \item Report bugs at \url{https://github.com/tor-gu/njoaguof/issues}
}

}
\author{
\strong{Maintainer}: Tor Gunston \email{tor.gunston@protonmail.com}

Authors:
\itemize{
  \item Tor Gunston \email{tor.gunston@protonmail.com}
}

}
\keyword{internal}
