% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_name_variants.R
\docType{data}
\name{officer_name_variants}
\alias{officer_name_variants}
\title{Officer name variants.}
\format{
A dataframe with 2 columns
\describe{
\item{officer_name_id}{Officer name id}
\item{officer_name}{Officer name variant}
}
}
\source{
\url{https://www.njoag.gov/force/}
}
\usage{
officer_name_variants
}
\description{
Every officer name and name id occurring in the raw data.
}
\details{
This data is extracted from the \code{officer_name} and \code{Officer_Name2}
fields in \code{use_of_force_raw}.

The raw data contains an officer name id and an officer name for each
incident. The officer name will sometimes vary from incident to incident in
spelling and capitalization. This table records every variant seen.

If you need to associate a specific name variant with a specific incident,
refer to the raw data.
}
\keyword{datasets}
