% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_of_force_raw.R
\docType{data}
\name{use_of_force_raw}
\alias{use_of_force_raw}
\title{Raw use of force data from the NJ OAG}
\format{
A dataframe with 51 columns
}
\source{
\url{https://www.njoag.gov/force/}
}
\usage{
use_of_force_raw
}
\description{
Use of force data downloaded from the NJ OAG Use of Force dashboard.
}
\keyword{datasets}
